feat(ActionsMenuButton): Add component
we could pass DotsIcon as prop but for now it's not necessary
