+++ react/ActionsMenu/Actions/locales/en.json
@@ -1,4 +1,5 @@
 {
+  "menu": "Menu",
   "viewInContacts": "View in Cozy Contacts",
   "viewInDrive": "Open",
   "modify": "Modify",
