+++ react/ActionsMenu/Actions/locales/fr.json
@@ -1,4 +1,5 @@
 {
+  "menu": "Menu",
   "viewInContacts": "Voir dans Cozy Contacts",
   "viewInDrive": "Ouvrir",
   "modify": "Modifier",
