+++ react/ContactsList/locales/en.json
@@ -1,6 +1,5 @@
 {
   "empty": "EMPTY",
   "me": "me",
-  "favorite": "favorites",
-  "menu": "Menu"
+  "favorite": "favorites"
 }
