+++ react/ContactsList/locales/fr.json
@@ -1,6 +1,5 @@
 {
   "empty": "VIDE",
   "me": "moi",
-  "favorite": "favoris",
-  "menu": "Menu"
+  "favorite": "favoris"
 }
